Recursive implementation/test for binary search

+++ src/main/java/com/coffeebeans/search/binary/BinarySearch.java
@@ -1,5 +1,5 @@
 package com.coffeebeans.search.binary;
 
 public interface BinarySearch<T> {
-    long search(final T t, final T[] values);
+    int search(final T t, final T[] values);
 }
